forgot to delete two remaining files that are no longer used. also moved rosafactory into its own service package and updated the tests as needed to reflect the new package structure.
