+++ src/main/java/org/praekelt/restforms/core/services/rosa/RosaFactory.java
@@ -1,4 +1,4 @@
-package org.praekelt.restforms.core.services;
+package org.praekelt.restforms.core.services.rosa;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
