+++ src/test/java/org/praekelt/restforms/core/services/RosaFactoryJedisClientIntegrationTest.java
@@ -1,5 +1,6 @@
 package org.praekelt.restforms.core.services;
 
+import org.praekelt.restforms.core.services.rosa.RosaFactory;
 import static org.junit.Assert.*;
 import org.junit.After;
 import org.junit.AfterClass;
